Added button of share book review via Twitter, changed styles and icons
