+++ pages/MainPage.vue
@@ -1,12 +1,10 @@
 <template>
   <NuxtLayout class="h-screen bg-white w-full">
     <div class="text-center w-full pl-60 pt-10">
-      <h2 class="font-semibold text-xl">{{ $t('welcome') }}</h2>
+      <h2 class="font-semibold text-xl">{{ $t("welcome") }}</h2>
     </div>
-    <div class="w-11/12 ml-96 m-auto">
-    </div>
+    <div class="w-11/12 ml-96 m-auto"></div>
   </NuxtLayout>
 </template>
 
-<script setup>
-</script>
+<script setup></script>
